Merge community's pull request - init
Momo is an experimental new digital currency that enables anonymous, instant payments to anyone, anywhere in the world. Momo uses peer-to-peer technology to operate with no central authority: managing transactions and issuing money are carried out collectively by the network. Momo Core is the name of the open source software which enables the use of this currency.